# docs/contracts/ordering.openapi.json
{
  "openapi": "3.0.3",
  "info": { "title": "ordering learning API", "version": "1.0.0" },
...
  "paths": {
...
    "/api/v1/products": {
      "get": {
        "responses": {
          "2XX": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "format": "uuid", "type": "string" },
                          "createdAt": { "format": "date-time", "type": "string" },
                          "updatedAt": { "format": "date-time", "type": "string" },
                          "name": { "type": "string" },
                          "description": { "type": "string" },
                          "priceCents": { "minimum": 0, "type": "integer" },
                          "availableStock": { "minimum": 0, "type": "integer" },
                          "active": { "type": "boolean" },
                          "deactivatedAt": {
                            "format": "date-time",
                            "type": "string",
                            "nullable": true
                          }
                        },
                        "required": [
                          "id",
                          "createdAt",
                          "updatedAt",
                          "name",
                          "description",
                          "priceCents",
                          "availableStock",
                          "active",
                          "deactivatedAt"
                        ]
                      }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "requestId": { "format": "uuid", "type": "string" },
                        "correlationId": { "format": "uuid", "type": "string" },
                        "respondedAt": { "format": "date-time", "type": "string" }
                      },
                      "required": ["requestId", "correlationId", "respondedAt"]
                    }
                  },
                  "required": ["data", "meta"]
                }
              }
            }
          }
        }
      },
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "type": "object",
                "properties": {
                  "name": { "minLength": 1, "maxLength": 120, "type": "string" },
                  "description": { "maxLength": 1000, "type": "string" },
                  "priceCents": { "minimum": 0, "maximum": 1000000000, "type": "integer" },
                  "availableStock": { "minimum": 0, "maximum": 1000000, "type": "integer" }
                },
                "required": ["name", "description", "priceCents", "availableStock"]
              }
            }
          }
        },
        "responses": {
          "2XX": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": { "format": "uuid", "type": "string" },
                        "createdAt": { "format": "date-time", "type": "string" },
                        "updatedAt": { "format": "date-time", "type": "string" },
                        "name": { "type": "string" },
                        "description": { "type": "string" },
                        "priceCents": { "minimum": 0, "type": "integer" },
                        "availableStock": { "minimum": 0, "type": "integer" },
                        "active": { "type": "boolean" },
                        "deactivatedAt": {
                          "format": "date-time",
                          "type": "string",
                          "nullable": true
                        }
                      },
                      "required": [
                        "id",
                        "createdAt",
                        "updatedAt",
                        "name",
                        "description",
                        "priceCents",
                        "availableStock",
                        "active",
                        "deactivatedAt"
                      ]
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "requestId": { "format": "uuid", "type": "string" },
                        "correlationId": { "format": "uuid", "type": "string" },
                        "respondedAt": { "format": "date-time", "type": "string" }
                      },
                      "required": ["requestId", "correlationId", "respondedAt"]
                    }
                  },
                  "required": ["data", "meta"]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/products/{id}": {
      "get": {
        "parameters": [
          {
            "schema": { "format": "uuid", "type": "string" },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "2XX": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": { "format": "uuid", "type": "string" },
                        "createdAt": { "format": "date-time", "type": "string" },
                        "updatedAt": { "format": "date-time", "type": "string" },
                        "name": { "type": "string" },
                        "description": { "type": "string" },
                        "priceCents": { "minimum": 0, "type": "integer" },
                        "availableStock": { "minimum": 0, "type": "integer" },
                        "active": { "type": "boolean" },
                        "deactivatedAt": {
                          "format": "date-time",
                          "type": "string",
                          "nullable": true
                        }
                      },
                      "required": [
                        "id",
                        "createdAt",
                        "updatedAt",
                        "name",
                        "description",
                        "priceCents",
                        "availableStock",
                        "active",
                        "deactivatedAt"
                      ]
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "requestId": { "format": "uuid", "type": "string" },
                        "correlationId": { "format": "uuid", "type": "string" },
                        "respondedAt": { "format": "date-time", "type": "string" }
                      },
                      "required": ["requestId", "correlationId", "respondedAt"]
                    }
                  },
                  "required": ["data", "meta"]
                }
              }
            }
          }
        }
      },
      "patch": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "type": "object",
                "properties": {
                  "name": { "minLength": 1, "maxLength": 120, "type": "string" },
                  "description": { "maxLength": 1000, "type": "string" },
                  "priceCents": { "minimum": 0, "maximum": 1000000000, "type": "integer" }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": { "format": "uuid", "type": "string" },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "2XX": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": { "format": "uuid", "type": "string" },
                        "createdAt": { "format": "date-time", "type": "string" },
                        "updatedAt": { "format": "date-time", "type": "string" },
                        "name": { "type": "string" },
                        "description": { "type": "string" },
                        "priceCents": { "minimum": 0, "type": "integer" },
                        "availableStock": { "minimum": 0, "type": "integer" },
                        "active": { "type": "boolean" },
                        "deactivatedAt": {
                          "format": "date-time",
                          "type": "string",
                          "nullable": true
                        }
                      },
                      "required": [
                        "id",
                        "createdAt",
                        "updatedAt",
                        "name",
                        "description",
                        "priceCents",
                        "availableStock",
                        "active",
                        "deactivatedAt"
                      ]
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "requestId": { "format": "uuid", "type": "string" },
                        "correlationId": { "format": "uuid", "type": "string" },
                        "respondedAt": { "format": "date-time", "type": "string" }
                      },
                      "required": ["requestId", "correlationId", "respondedAt"]
                    }
                  },
                  "required": ["data", "meta"]
                }
              }
            }
          }
        }
      },
      "delete": {
        "parameters": [
          {
            "schema": { "format": "uuid", "type": "string" },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "2XX": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": { "format": "uuid", "type": "string" },
                        "createdAt": { "format": "date-time", "type": "string" },
                        "updatedAt": { "format": "date-time", "type": "string" },
                        "name": { "type": "string" },
                        "description": { "type": "string" },
                        "priceCents": { "minimum": 0, "type": "integer" },
                        "availableStock": { "minimum": 0, "type": "integer" },
                        "active": { "type": "boolean" },
                        "deactivatedAt": {
                          "format": "date-time",
                          "type": "string",
                          "nullable": true
                        }
                      },
                      "required": [
                        "id",
                        "createdAt",
                        "updatedAt",
                        "name",
                        "description",
                        "priceCents",
                        "availableStock",
                        "active",
                        "deactivatedAt"
                      ]
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "requestId": { "format": "uuid", "type": "string" },
                        "correlationId": { "format": "uuid", "type": "string" },
                        "respondedAt": { "format": "date-time", "type": "string" }
                      },
                      "required": ["requestId", "correlationId", "respondedAt"]
                    }
                  },
                  "required": ["data", "meta"]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/products/{id}/stock": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "type": "object",
                "properties": {
                  "delta": { "minimum": -1000000, "maximum": 1000000, "type": "integer" }
                },
                "required": ["delta"]
              }
            }
          }
        },
        "parameters": [
          {
            "schema": { "format": "uuid", "type": "string" },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "2XX": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": { "format": "uuid", "type": "string" },
                        "createdAt": { "format": "date-time", "type": "string" },
                        "updatedAt": { "format": "date-time", "type": "string" },
                        "name": { "type": "string" },
                        "description": { "type": "string" },
                        "priceCents": { "minimum": 0, "type": "integer" },
                        "availableStock": { "minimum": 0, "type": "integer" },
                        "active": { "type": "boolean" },
                        "deactivatedAt": {
                          "format": "date-time",
                          "type": "string",
                          "nullable": true
                        }
                      },
                      "required": [
                        "id",
                        "createdAt",
                        "updatedAt",
                        "name",
                        "description",
                        "priceCents",
                        "availableStock",
                        "active",
                        "deactivatedAt"
                      ]
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "requestId": { "format": "uuid", "type": "string" },
                        "correlationId": { "format": "uuid", "type": "string" },
                        "respondedAt": { "format": "date-time", "type": "string" }
                      },
                      "required": ["requestId", "correlationId", "respondedAt"]
                    }
                  },
                  "required": ["data", "meta"]
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/cache/actions": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "type": "object",
                "properties": {
                  "action": {
                    "anyOf": [
                      { "type": "string", "enum": ["clear"] },
                      { "type": "string", "enum": ["expire"] },
                      { "type": "string", "enum": ["corrupt"] }
                    ]
                  }
                },
                "required": ["action"]
              }
            }
          }
        },
        "responses": {
          "2XX": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {},
                    "meta": {
                      "type": "object",
                      "properties": {
                        "requestId": { "format": "uuid", "type": "string" },
                        "correlationId": { "format": "uuid", "type": "string" },
                        "respondedAt": { "format": "date-time", "type": "string" }
                      },
                      "required": ["requestId", "correlationId", "respondedAt"]
                    }
                  },
                  "required": ["data", "meta"]
                }
              }
            }
          }
        }
      }
    }
  }
}